src: check compiler page was given the expected PROT_FLAGS
in HardenedBSD[1] and probably other security enhanced platforms, the
PROT_FLAGS passed to mmap won't be respected in some cases, leading
to the allocation of pages that lack execution permission and therefore
will trigger the application to segfault later.

detect the problem earlier by doing an mprotect of the recently allocated
page with the same permissions and if denied, discard it instead and
report an allocation problem upstream.

[1] https://github.com/HardenedBSD/hardenedBSD/wiki#pageexec-and-mprotect-aka-noexec