protexec: use memfd_create for file backing creation
memfd_create is available in linux since kernel 3.17 with support
added in glibc (2.27), musl (1.1.20) and bionic (30).

it allows creating a tmpfs backed file without having to worry about
file system support, path location, permission or even file name
so is ideal for the kind of file that the ProtExecutionAllocator
needs to back its maps.

this solves the problems reported in PCRE's bug 2455[1] where a JIT
enabled call that was compiled to use this experimental allocator
will fail because the selected TMPDIR was non executable (ironically
something common in SE systems)

[1] https://bugs.exim.org/show_bug.cgi?id=2455