protexec: restrict file system permissions in the map files
the backing file contents should not be accessible outside the running
process so make sure there are no RW permissions in the file created
as early as possible.

memfd_create will create the file as 0600, and while it is not linked
anywhere in the filesystem is still accessible through its proc link
in /proc/$PID/fd/$FD; after this change that link won't be accessible
outside the running process.

mkostemp will create the file as 0600 most of the time, but since
POSIX mkstemp doesn't mention a mode it is safer to always make sure
a fitting umask is used or if that is not possible that its mode is
changed ASAP so it won't be accessible outside the running process.

note that for multithreaded users of this code, there is a race
condition that could still allow access to the underlying file
backed files and therefore using the regular allocator will be
safer.