config: disable global lock by default
it is not used anywhere and is unlikely to be useful so discourage
its use by requiring a compile time flag to enable.

since sljit is already documented to be thread safe, and the allocators
(which are not, with the exception of "wxexec") already have their own
lock, and it is known to be unsafe in windows, there is likely nothing
to lose, so document its obsolescence with a timeframe to removal.