utils: use posix_madvise(MADV_DONTNEED) to reclaim stack memory
after it was reported that madvise was no longer defined in Solaris
(ex: if using CFLAGS="-D_XOPEN_SOURCE=600") and was never available
for Haiku, with a failed tentative fix committed earlier.

posix_madvise is part of POSIX and therefore well supported[1], but
its implementation of MADV_DONTNEED is not exactly equivalent (at
least on Linux) which will be addressed in the next commit.

[1] https://www.gnu.org/software/gnulib/manual/html_node/posix_005fmadvise.html

Fixes: bdc4d475f277cd8abefb493484201e3addf3c591