protexec: drop support of NetBSD < 8
NetBSD-7 is EOL as of June 29th 2020, so assume all NetBSD systems will use
their system specific alloc_chunck().

This breaks building with "unsupported" versions and any supported version
that might not define the NetBSD extensions used by it, so add a comment
to explain that.

Since it is no longer needed, remove the ugly compatibility layer for
secure_getent, that was used in the Linux specific part to support now
obsoleted versions of NetBSD.

Suggested-by: @krytarowski