wxexec: fail early if WX pages are forbidden per policy
in security enhanced systems like SELinux or PaX, the use of WX
pages might be forbidden by policy (unless exceptions are added).

check the policy or try to infer its settings through mprotect
and prevent a SEGFAULT later.

NetBSD with PROT_MPROTECT is excempted from the test, which will
be cached for all others to reduce the impact on performance.