util: remove global lock
as shown in the long running bug for the windows version of the
executable allocator lock and its fixed implementation, it is
difficult to get it right at this level.

if there is a need to lock sljit or parts of it, it will be better
that it will be done at the application layer.