windows: fix race for initialization of executable allocator lock
the locks (from utils) are not thread safe in Windows, because they
have no way to synchronize initialization between threads, and therefore
could have more than one thread initializing the handler to different
values.

for compatibility with obsolete versions of windows (and ReactOS)
implement a "poor man" lock for the handler using the Interlocked API
to synchronize a local handler with the global one and discard the
extra ones.

Helped-by: Giuseppe D'Angelo <dangelog@gmail.com>