test: big refactoring in preparation to add windows thread support
in preparation to being able to run tests as threads, change their
declaration to be compatible with the one used on each platform and
use the return value to signal success as a bool.

this allows to reduce the need of individual tests to change directly
the global sucess tests counter, so add some macros to help  abstract
the platform specific variance further..

FAILED() conflicts with a macro of the same name in windows so rename
it; most of the changes are because of this.