test: teach sljit_test to run tests in threads (WIP)
add a commandline to instruct it to run each test multiple times in
different threaads, with '0' meaning not to use threads and the
default changing to use 1 worker thread.

this change of default behaviour is a sideffect of the current
limitations with command line processing and so is likely to be
corrected later; the same applies to not being able to do verbose
unless using the default number of threads.

test64 is unstable (as reported independently)