test64: hack allocation so it will reuse address
change the executable memory allocation function to a local one that
allows returning the same allocation twice (or more) times by flagging
it through a global flag.

since it requires a global variable is not thread safe, but at least
it is no longer dependent on internal implementation details of the
executable allocator used.

while at it, document better the test and constrains and avoid using
a potentially invalid pointer by deferring the deallocation of the
code to the end of the test by the standard code.code alias.