config: allow partial overriding of executable allocator
test64 has special needs for the addresses provided by the allocator
but can't use a custom one without affecting other tests.

create a middle ground that allows for a custom sljitLir to be created
pointing to an externally provided function that will be called for
allocation of executable memory.

to redefine to "function" pass -DSLJIT_MALLOC_EXEC=function and make
sure the function definition matches the interface and it is globally
visible.

the original function will be available for fallback using its known
name: sljit_malloc_exec().