macos: allow runtime detection if target is older than BigSur
instead of aborting the build, allow for runtime detection of the
API needed for JIT in Apple Silicon and which was released with
macOS 11.

since that was also the first release to support Apple Silicon the
runtime check should never be needed in practice and other than a
small performance impact its main effect is to allow universal
builds that require an older version of the OS as a minimum.