sparc: avoid NOP implementation for clear_cache with gcc (#117)
* sparc: avoid NOP implementation for clear_cache with gcc

The fallback code was masked after 0863c54 (Support Android ARM64
cacheflush. Patch by Tavian Barnes., 2016-04-06), so move it up.

Shown breaking test 12 in Solaris 10/11 with gcc <= 5.5.0

* sparc: avoid NOP implementation for clear_cache builtin

Shown breaking test12 in Debian 12 with gcc >= 10 (which is the first
to define __has_builtin; gcc-9 also fails if using the builtin but
will use the fallback).