build: add barebones cmake for Windows
Building sljit requires GNU make, and uses a syntax that can be
easily converted into BSD or even NMAKE if needed, but in Windows
users could be tempted to try MSys or MinGW as a way to get make
working, only to find a broken library.

Provide instead with a cmake recipe that could be used to generate
whatever is that they use to build with native tools, and so be
able to test or improve sljit more easily.

This script is intentionally minimal to avoid duplicating the logic
from the makefile and it even it to avoid hardcoding values that
are relevant, but to avoid accidentally overwriting that file, as
well as making it clear that it uses GNU syntax had it renamed.

While at it, increase the Windows stack size and fix some spelling
issues reported by MSVC once it was able to load sljit as a project.