wxexec: officially drop support for EOL NetBSD < 8 versions
f6add53 (wxexec: better support for SE systems including NetBSD (#67),
2020-08-26) adds NetBSD specific code to improve its usability, but
was not careful enough to make sure the facilities required where
available in older versions and as a result it fails to build with
NetBSD < 8.

Remove the code to support those old systems and as a side effect
break the build if the development environment doesn't have support
for the required extensions in newer versions (ex: -D_XOPEN_SOURCE)

This is consistent with the behaviour of the "protexec" allocator
and since it will require system/executable exceptions to the policy
just like the default allocator is an alternative solution that is
unnecessarily inefficient and therefore its use should be discouraged.