windows: allow building with MinGW64 gcc/clang
Currrently showing errors like the following:

  cc -Isljit_src -DSLJIT_HAVE_CONFIG_PRE=1 -Itest_src -O2 -Wall -Wextra
  -Wconversion -Wsign-compare -Werror  bin/sljitMain.o
  test_src/sljitTest.c sljit_src/sljitLir.c -o bin/sljit_test -lm -lpthread
  sljit_src/sljitLir.c: In function 'sljit_verbose_param':
  sljit_src/sljitLir.c:945:44: error: format '%d' expects argument of type
  'int', but argument 3 has type 'sljit_sw' {aka 'long long int'}
  [-Werror=format=]
    945 |                 fprintf(compiler->verbose, "#%" SLJIT_PRINT_D
        "d", (i));
        |                                            ^~~~
  ~~~
        |
  |
        |
  sljit_sw {aka long long int}

Add a check for __GNUC__ that would be used for both gcc and clang to
set the int64_t print format to the non MSVC standard in 64bit.

Alternatively, migrating to the C99 PRIu64 macros was punted for now
but might be worth considering later.