windows: fix building in arm64 with clang
Failing at least with Visual Studio 2022 when using the clang
compiler with:

  lld-link : error : undefined symbol: __clear_cache

Reorder the checks so that the Windows specific call is used
before the compiler builtin.

Might need adjusting when/if clang or gcc from MinGW also add
support for that architecture in that platform and also if
clang is expanded to support the 32bit version.