arm: add support for detecting SIMD (NEON) availability (WIP)
Requires getauxval() which is available at least in Linux/Android
with recent versions of the libc and that is therefore behind a
configure like macro.

A similar function from FreeBSD>=12, Windows and NetBSD is used
for each case.

While at it, consolidate the code to use the same externally
visible flag that is used in x86 for SSE2.