riscv: disable FPU when using soft-float ABI
FPU support is optional and therefore is missing for ABI: ilp32 or lp64

The possibility of targeting a single float abi (ex: ilp32f) is punted
from the proposed logic as it is unlikely to be used outside of embedded
and will require further intrusive changes.

Changes to fully support soft-float also pending.