arm: refactor test92 to avoid possibly conflicting writes
The specification allow up to a maximum exclusive granule of 512 words
to be reserved by `ldxr`, so writes close enough to the address that
was loaded could result on the reservation being invalidated and `stxr`
failing.

Instead of writing back the values, do a second normal read to validate
the first one and compare them internally.  While at it make the test
logic slightly easier to read.

Fixes: #160