loongarch: sljit_emit_atomic_load/store implementation
Assumes all atomics are at least 32bit aligned by setting
SLJIT_ATOMIC_WIDTH.

In that mode, only the bits that correspond to the operand
size are used and the rest are pressumed to be 0.

Operations with lower alignment than it are undefined and
will result in a SIGBUS.

If SLJIT_ATOMIC_EMULATION is defined, then additional code
is enabled that allow manipulating those narrower types at
a significant performance cost.

Original code by Mingtao, all bugs mine.