arm: restore emulation for SLJIT_CTZ in armv6
Since 3499aba (Set armv6 as the minimum arm version, 2023-04-07)
the fallback code in use for ARMv5 was removed, but the RBIT
instruction is not supported in ARMv6 so bring it back.

A similar problem in the Thumb2 code is suspected as it should
affect at least armv6t2 but a solution for it has been punted.