build: clear warnings with gcc12 and improve Makefile
Mostly fix -Wsign-conversion warnings with gcc12 affecting at least
riscv and power when using sanitizers.

To help with linking "-latomic" when needed, add a EXTRA_LIBS
variable and while at it add a few more that can be used to set
the optimization flags or disable -Werror.

While at it, do some refactoring and cleanup for added warnings.