ppc: correctly implement SLJIT_REV in 64bit little endian
Otherwise failing cases 3 and 5 in test87 with swapped words

Add code to load and store the 2 integers while byte swapping
them in an endianess aware way.

While at it, add native support for 64bit swaps that could be used
as an alternative in POWER7 and POWER10 and set SLJIT_HAS_REV as
not being provided by emulation for the later.