mips: correctly detect valid FPU and support StatusFR=1 in 32bit
The original detection only considers an FPU valid if it has 64bit
registers, but the code allows for doubles stored in a pair of
32bit registers and actually assumes that is always the case.

Add individual feature bits for each option and enable the FPU if
it supports doubles, regardless of register size for MIPS32, which
should at least add support for the original MIPS32 and eve older
CPUs.

While technically not a CPU configuration, track the StatusFR flag
to allow for supporting true 64bit registers, but protect that
detection behind an optional SLJIT_DETECT_FR define.

Starting with MIPSr2, add code to support StatusFR=1 and which can
be used on either mode and enable it by default after MIPSr5,
where is no longer optional.

While at it, improve autodetection and do some refactoring.