mips: allow detection of StatusFR with prctl()
Use the Linux (since kernel 4.0) for MIPS interface to prctl() to
detect which o32 ABI variant is in effect for FPU register sizes
and prefer this method if available.

Keep the original check for a fallback but make it only available
when SLJIT_DETECT_FR == 2, unless it is the only option.

While at it, do some refactoring to make sure StatusFR is always
set.