loongarch: fix FTBFS when building pcre2test as part of PCRE2
When `sljit_u32 cpu_feature_list` was introduced, it wasn't set as
static and therefore would fail linking if multiple copies of it
were used as is done by pcre2test.

Make it static (as used everywhere) and while at it, rename the
`dst` parameter in `sljit_emit_fop2r` as well as other styling
changes for consistency.