build: make code mostly ANSI C89 compatible
Add a variable that can be used to enforcce compatibility to a
standard and enable it for C89, but allow it to be overriden
for cases where the code is using C99 (s390x) or requires not
compatible features (powerpc64 with trampolines).

While at it, update some of the examples so they will build.