x86: improve SIMD detection
Since 56dbde0 (Add xgetbv feature detection support on x86, 2024-03-23),
AVX support can be "rescinded" if it has been partially disabled, but
the check added didn't include the possibility that OSXSAVE could be
also disabled (ex: using the `noxsave` kernel parameter in Linux).

Correct that and while at it, add a similar check for SSE2 (only relevant
in 32bit) and for the possibility that if not enabled (ex: the OS never
set CR4.OSFXSR), then SSE41 support should also be rescinded.