misc: avoid unnecesary shadowed variable
When building with the default allocator that uses a global named
`total_size`, uses of a similar variable mame elsewhere will
trigger warnings.

Rename the local variable to something more accurate and make sure
that by default the warning that will trigger otherwise (supported by
gcc and clang) is added to cover regressions.