config: add SLJIT_FALLTHROUGH to exempt -Wimplicit-fallthrough
When the flag was introduced (clang++ 3.2 and gcc 7), comments that are
detected by gcc as hints to suppress this alert were added to the codebase,
but those were never recognized by clang.

This triggers warnings when building with clang and using that flag,
which could be prevented by moving to an attribute instead.

Make sure to target only gcc/clang or derivatives that will need
this new macro (icc does not support -Wimplicit-fallthrough), to avoid
unnecessary incompatibilities and to provide an empty fallback for all
other versions and compilers.

Suggested by: Gernot Gebhard <gebhard@absint.com>