config: add SLJIT_FALLTHROUGH to exempt -Wimplicit-fallthrough or C5262 (#330)
When the flag was introduced (clang++ 3.2 and gcc 7), comments that are
detected by gcc as hints to suppress this alert were added to the codebase,
but those were never recognized by clang. MSVC 17.4 (and SAL before that)
introduced C5262 with the same objective but also their own syntax.

This triggers warnings when building with clang and using that flag,
which could be prevented by moving to an attribute instead. Both C23 and
C++17 provide a standard one, so add support for those and extra rules to
accomodate the older syntax in GNU compilers.

Make sure to target only gcc/clang or derivatives that will need this new
macro (icc, nvc, bcc32 do not support -Wimplicit-fallthrough), to avoid
unnecessary incompatibilities when using GNU attributes; this removes
clang C compilers 3.2 to 3.5 which trigger -Wmissing-declaration warnings
in some configurations when the fallback attribute is otherwise supported
and clang < 6 when have_c_attribute is not, and provide an empty fallback
for all others.

Suggested by: Gernot Gebhard <gebhard@absint.com>